feat: implement CLI and command skeleton
Commands are largely hollow due to the fact that this is just the first
draft, and to allow myself to add everything as we go along.
I have also settled on skipping "real" CLI output - we can just adjust
the logger to format the strings differently for an shell.
This way, somebody can implement a GUI later on and the logging can stay
the same.