feat: implement a config manager
As many classes would need access to the configured working_dir (and
later on surely other vars), the need for a ApplicationConfigManager
arised. As this is a cross-cutting-concern, having a Singleton that can
just be called wherever needed seemed more appropriate than passing the
vars around as needed - at last for now.
Normally, I would like all classes to only have access to the params
they actually need, but I feel silly to write "working_dir" as a
argument to every class constructor. It is a cross cutting concern. I
don't pass around the logger either. A better way would be creating
annotations or injecting the vars directly, but these solutions would be
similarly in complexity.