feat: implement cache for downloading different Moodle versions
As we want to test a Boost Union version/commit/PR in all supported
Moodle versions, we need to download said Moodles. To decrease the
pressure on the traffic, we just download them as .tar to disk and hit
the disk first to check if we already have dl'd selected Moodle version.
If not, we had a cache miss and need to ask the far, far away GitHub
server to please provide us with the correct file.