feat: implement MoodleDownloader to handle exceptional HTTP error cases
As I developed this application, GitHub had several small hiccups which
led to me thinking I had a bug or logical error in my programming when
instead I just got a random timeout.
A better downloader that performs a few retries and correctly prints the
error should greatly increase the usability and reliability.

Added benefit:
We now handle the case correctly if the user tries to create a test env
for a Moodle version that does not exist.