fix: Moodle DB tables being empty when the DB container wasn't ready.
As we didn't wait for the DB container to be started, before we started
the call to configure the environment for manual testing, an issue could
that led the user to believe that the operation was complete; but we
actually had default data for the admin accounts - leading to users not
being able to log in.

Remedy this by waiting for the DB to start.