feat: redesign reverse proxying to use a sub-location instead of a subdomain
Due to the fact that we do have a wildcard certificate for our
technical domain at Plesk and cannot get one due to an issue with it's
DNS, we will rather use sub-locations.
The aforementioned issue is the following: Plesk's DNS does not
propagate TXT entries needed for the ACME-DNS-challenge from Let's Encrypt.
Therefore the only way to get one is to pay money, which isn't an
option.