feat: implement system to choose the correct php image
To make sure that a Moodle container will be run with it's correct PHP
version, a system is needed to choose the correct docker image tag as
said tag denotes the PHP version used inside said image.
This is needed because we want to make sure that the Moodle container we
create can actually be used. The moodle-docker project defaults to PHP
version 8.1, which would mean in case we want to create a Moodle
Container with version 4.0.0, it wouldn't start as the newest PHP
version this Moodle supports is 8.0.

Will be selected by checking which is the newest, supported PHP version
of the given Moodle version.